+++ plugins/account_history_api_plugin/account_history_api_plugin.cpp
@@ -18,12 +18,6 @@ account_history_api_plugin::~account_history_api_plugin(){}
 void account_history_api_plugin::set_program_options(options_description&, options_description&) {}
 void account_history_api_plugin::plugin_initialize(const variables_map&) {}
 
-struct error_results {
-  uint16_t code;
-  string message;
-  string details;
-};
-
 #define CALL(api_name, api_handle, api_namespace, call_name) \
 {std::string("/v1/" #api_name "/" #call_name), \
    [this, api_handle](string, string body, url_response_callback cb) mutable { \
@@ -61,5 +55,3 @@ void account_history_api_plugin::plugin_startup() {
 void account_history_api_plugin::plugin_shutdown() {}
 
 }
-
-FC_REFLECT(eos::error_results, (code)(message)(details))
