+++ plugins/chain_api_plugin/chain_api_plugin.cpp
@@ -26,12 +26,6 @@ chain_api_plugin::~chain_api_plugin(){}
 void chain_api_plugin::set_program_options(options_description&, options_description&) {}
 void chain_api_plugin::plugin_initialize(const variables_map&) {}
 
-struct error_results {
-  uint16_t code;
-  string message;
-  string details;
-};
-
 #define CALL(api_name, api_handle, api_namespace, call_name) \
 {std::string("/v1/" #api_name "/" #call_name), \
    [this, api_handle](string, string body, url_response_callback cb) mutable { \
@@ -86,5 +80,3 @@ void chain_api_plugin::plugin_startup() {
 void chain_api_plugin::plugin_shutdown() {}
 
 }
-
-FC_REFLECT(eos::error_results, (code)(message)(details))
