+++ plugins/http_plugin/include/eos/http_plugin/http_plugin.hpp
@@ -5,6 +5,8 @@
 #pragma once
 #include <appbase/application.hpp>
 
+#include <fc/reflect/reflect.hpp>
+
 namespace eos {
    using namespace appbase;
 
@@ -74,4 +76,14 @@ namespace eos {
         std::unique_ptr<class http_plugin_impl> my;
    };
 
+   /**
+    * @brief Structure used to create JSON error responses
+    */
+   struct error_results {
+     uint16_t code;
+     string message;
+     string details;
+   };
 }
+
+FC_REFLECT(eos::error_results, (code)(message)(details))
